moved spy.js up one with promise spy for client
related #25
